Creates BasicBeingTestState file
This is a class that extends basic state and will be used for the testing
of the basic being class. In this way the Basic being can be updated and
rendered as if in the Production PlayState. The render method prints
a hellow test Basic being string + n. "n" is a variable of this test state
that is being incremented in teh update method.