Completes issue 31 checkout issue32-basic being world position vector.
The Basic Being currently only has a vector denoting the position in screen space. For client controlled hero this will be static in screen space where as. Everything else will be relative to the position of the client on the screen. World Position is the Being location on the world map. These two things are distinct I believe.