Creates a boolean isClient variable in BasicBeing
this flag can be set when a client so that the screen position of the client Beings will not be updated. Beings who are not client must have their screen position affected by the position of the client. because as they move around the map the location of "screen 0,0" is translating.